Use more sophisticated 'cpanfile' format in Git
To allow installing Coocook from a Git checkout we now use
'cpanfile' in the root directory. It is autogenerated from `dzil build`.
It is supported by cpanminus and is a more sophisticated format. See

http://blogs.perl.org/users/mark_allen/2013/07/why-i-use-cpanfile-and-you-should-too.html

The traditional Makefile.PL is still contained in the build directory if
you add Coocook-*.tar.gz to your private CPAN mirror or we ever upload
Coocook to official CPAN.

To install dependencies from Git still just do

    cpanm --installdeps .