Add role_relation to config of Authentication::Store::DBIx::Class
We recently got HTTP 500 errors on Coocook.org for

    GET https://coocook.org/user/${ANY_USERNAME}

because

    Couldn't render template "user/show.tt:
    undef error - user->roles accessed, but no role configuration found"

I could not reproduce this error in my development environment but
it looks like Catalyst::Authentication::Store::DBIx::Class::User
fetches the roles if not prefetched. Obviously this never happened
before so I never needed to actually configure where the roles are
found.

Unfortunately this duplicates the logic in
Coocook::Schema::Result::User->roles().

I guess $user->roles() is not called because $c->user is actually
a wrapper object around a Result::User object.