Fix multiple copies of shop sections in purchase list
Actual problem was complex SQL query without {distinct => 1}.

To test the issue this commit expands the purchase list in
the test data and adjust the tests accordingly.

In t/purchase_lists.t I use more tools from Test::Deep to
minimize the size of the data structure needed in the test file.