Fix ResultSet::Project->inventory() to run counts only 1x
I noticed--thanks to a warning issued from ResultSet->single()--
that the previously generated SQL ran all the counts multiple times.

Before search() generated something like this which rans the COUNTs
once for every project (!) existing:

    SELECT (COUNT * FROM articles WHERE project = 42) FROM projects;

Now it runs something like this:

    SELECT (COUNT * FROM articles WHERE project = 42) FROM projects
        WHERE projects.id = 42;

In reality the conditions in the subqueries are unnecessarily more
compilcated with JOINs but that needs more research.

I did some research on the outer wrapping

    SELECT ... FROM projects WHERE project.id = 42;

though and in SQLite it seems that this part doesn't do any harm
when running the queries 1000 times (measured with DBI::ProfileDumper).