Handle HTTP POST requests, replace ActionRole::RequireSSL
Catalyst::ActionRole::RequireSSL simply dies when a HTTP POST
request is received. Because the HTTPS check should happen
in the Controller::Root::base chain action it must handle
all kinds of HTTP methods.

I'm not yet sure what to do with HTTP POST requests. With
newer browsers it seems to be possible to redirect them
but what is the usecase? Links to HTTP can be posted anywhere
but POST request should originate from Coocook and all use
HTTPS anyway.

But we see a lot of HTTP POST requests on Coocook.org which
resulted in HTTP 500 until this commit.