Fix $c->user not raising exceptions, fix detected issues
I tested group permissions only as site admin before. Testing
as normal user raised a lot of issues.

Original cause is that $c->user is a wrapper around Result::User
with a bad AUTOLOAD method which hides exceptions for non-existing
methods being called.

This adds a workaround for that and solves these issues:

- typo 'rule' in method name

- wrong method name has_group_role()

- method has_any_group_rule() entirely missing

- role names wasn't available in TT stash if no more users to add