Fix deletion of projects in PostgreSQL
FK constraints need to be explicitly deferred until the end of
a transaction to pass.

I checked alternatively to set ON DELETE CASCADE for some
FK constraints but then we'd need to enable that for many
relationships that would normally make no sense to cascade.
I'd rather keep them as safety barriers.