Fix warning from t/postgresql.t if Pg is destroyed before DBIC
For a good example see this Travis CI log:
https://travis-ci.org/github/dboehmer/coocook/jobs/752858215#L829

    t/postgresql.t ................ 18/18 	(in cleanup) {UNKNOWN}: DBI Exception: DBD::Pg::st DESTROY failed: FATAL:  terminating connection due to administrator command
    server closed the connection unexpectedly
            This probably means the server terminated abnormally
            before or while processing the request. [for Statement "SELECT ...

On my system the warning looks a bit weird and was hard to debug.
I've probably shortened the life of my SSD by some time because
a loop of t/postgresql.t write over 100 MB every few seconds ...

There is a hidden warning that Pg server is unavailable. That
happens if destroy chooses to destroy Pg before DBIC.  That's
why the warning wasn't stable but appeared only in about 50%
of test runs.

This is how it looked, including the random string which might
contain some non-printable chars (probably random bytes):

    t/postgresql.t .. 13/14         (in cleanup) �e!nUpDnU at /home/daniel/perl5/perlbrew/perls/perl-5.32.0-threads-buster/lib/site_perl/5.32.0/DBIx/Class/Exception.pm line 77 during global destruction.
    t/postgresql.t .. ok
    All tests successful.
    Files=1, Tests=14, 26 wallclock secs ( 0.02 usr  0.00 sys +  5.12 cusr  2.11 csys =  7.25 CPU)
    Result: PASS