Set sane pragmas via use TestDB/Test::Coocook[::Base]
I wanted to fix a single warning because of a Unicdoe char in
some test output but I was tired of copying these pragmas
over and over again.

The warning came from the leftover debugging statement

    note $t->content

that printed the HTML with Unicode chars and caused the warning:

    Wide character in print at /home/daniel/perl5/perlbrew/perls/perl-5.32.0-threads-buster/lib/site_perl/5.32.0/Test2/Formatter/TAP.pm line 125.

The debugging statement is removed but Test2::Plugin::UTF8
fixes output of subtest names and test statements containing
Unicode chars.